src/requirements.txt
imageio==2.9.0
imageio-ffmpeg==0.4.3
matplotlib==3.4.1
numpy==1.19.5
tensorflow-gpu==2.4.1
torch==1.8.1
tqdm==4.60.0
